conversation_service/pyproject.toml
# This file uses the PEP 621 standard for project metadata.
[project]
name = "conversation-service"
version = "0.1.0"
description = "Real-time conversation service with WebSocket streaming and multi-agent hooks."
authors = [
    { name = "kgents", email = "dev@kgents.local" },
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  # Core web framework
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn[standard]>=0.30.1,<1.0.0",

  # Settings and data validation
  "pydantic[email]>=2.5.0,<3.0.0",
  "pydantic-settings>=2.3.4,<3.0.0",

  # Database
  "sqlalchemy[asyncio]>=2.0.31,<3.0.0",
  "psycopg[binary,pool]>=3.1.8,<4.0.0",
  "alembic>=1.13.1,<2.0.0",

  # Security (for validating JWTs from auth_service)
  "python-jose[cryptography]>=3.3.0,<4.0.0",

  # HTTP client for internal/external calls
  "httpx>=0.27.0,<1.0.0",
  # Rate limiting & middleware
  "slowapi>=0.1.9,<1.0.0",
  # Env loading for Alembic and bootstrap
  "python-dotenv>=1.0.1,<2.0.0",
]

[tool.poetry]
name = "conversation-service"
version = "0.1.0"
description = "Real-time conversation service with WebSocket streaming and multi-agent hooks."
authors = ["kgents <dev@kgents.local>"]
readme = "README.md"
packages = [{ include = "conversation_service", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic = { extras = ["email"], version = "^2.7.4" }
pydantic-settings = "^2.3.4"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.31" }
psycopg = { extras = ["binary", "pool"], version = "^3.1.19" }
alembic = "^1.13.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
httpx = "^0.27.0"
slowapi = "^0.1.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
asgi-lifespan = "^2.1.0"
httpx = "^0.27.0"

# Code Quality & Formatting
ruff = "^0.4.10"
black = "^24.4.2"
mypy = "^1.10.0"

# Development Workflow
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
